+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Changed
+- Improved error handling in ContainerInteropLoggerFactory.
+
 ### Fixed
 - Proper ordering of handlers and processors.
 - `ContainerInteropLoggerFactory` resolves logger configuration from either `Config` and `config` container entries.